requirements.txt
# UI Libraries
CustomTkinter

# macOS integration (for app tracking)
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
pyobjc-framework-ScriptingBridge

# LLM Interaction
llama-cpp-python

# Database
SQLAlchemy

# Plotting
matplotlib

# Other utilities
# psutil (for system information, might be useful)
Pillow # Pillow is needed for creating dummy images in tests and potentially for other image ops

# Packaging
PyInstaller 